+++ changelog.md
@@ -1,5 +1,37 @@
 # Michelle Changelog
 
+## 1.0.5, 20210328
+
+### Updated
+- Changing default font to Inter and size to 19px
+- Screenshot
+- Removing TGMPA script and plugin suggestions
+- Localization
+- Styles
+- Improving border color styles
+- Starter content
+- Block patterns
+
+### File updates
+	changelog.md
+	readme.txt
+	style.css
+	assets\scss\blocks.scss
+	assets\scss\comments.scss
+	assets\scss\content.scss
+	assets\scss\editor-style-blocks.scss
+	assets\scss\editor-style-classic.scss
+	assets\scss\global.scss
+	includes\Autoload.php
+	includes\Content\Starter.php
+	includes\Customize\Options.php
+	includes\Customize\RGBA.php
+	includes\Plugin\Component.php
+	languages/*.*
+	templates\parts\block\pattern-gallery-captions.php
+	templates\parts\block\pattern-team-2.php
+
+
 ## 1.0.4, 20210327
 
 ### Updated